# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/reservations/{id}/folio": {
      "get": {
        "operationId": "BillingController_getFolio",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accept-language",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Compte du séjour : lignes, encaissements, solde",
        "tags": [
          "Facturation"
        ]
      }
    },
    "/v1/reservations/{id}/folio/charges": {
      "post": {
        "operationId": "BillingController_addCharge",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accept-language",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChargeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Porter une prestation au compte du séjour",
        "tags": [
          "Facturation"
        ]
      }
    },
    "/v1/reservations/{id}/folio/charges/{chargeId}": {
      "delete": {
        "operationId": "BillingController_removeCharge",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chargeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accept-language",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Facturation"
        ]
      }
    },
    "/v1/rooms/{id}/charges": {
      "post": {
        "operationId": "BillingController_chargeRoom",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChargeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Facturation à la chambre : prestation portée au séjour en cours dans cette chambre",
        "tags": [
          "Facturation"
        ]
      }
    },
    "/v1/reservations/{id}/payments": {
      "post": {
        "operationId": "BillingController_pay",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accept-language",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Enregistrer un acompte, un paiement ou un remboursement",
        "tags": [
          "Facturation"
        ]
      }
    },
    "/v1/reservations/{id}/invoice": {
      "post": {
        "operationId": "BillingController_issue",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accept-language",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueInvoiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Émettre la facture du séjour (numérotation continue, pièce inaltérable)",
        "tags": [
          "Facturation"
        ]
      }
    },
    "/v1/invoices": {
      "get": {
        "operationId": "BillingController_list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Jeton `nextCursor` de la page précédente",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fields",
            "required": false,
            "in": "query",
            "description": "Champs à renvoyer, séparés par des virgules (ex. id,status,guest)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Numéro de pièce ou nom du client",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "invoice",
                "credit_note"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Facturation"
        ]
      }
    },
    "/v1/invoices/integrity": {
      "get": {
        "operationId": "BillingController_verify",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Vérifier empreintes, chaînage et continuité de la numérotation",
        "tags": [
          "Facturation"
        ]
      }
    },
    "/v1/invoices/{id}": {
      "get": {
        "operationId": "BillingController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Facturation"
        ]
      }
    },
    "/v1/invoices/{id}/credit-note": {
      "post": {
        "operationId": "BillingController_credit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreditNoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Annuler une facture par un avoir total",
        "tags": [
          "Facturation"
        ]
      }
    },
    "/v1/accounting/exports": {
      "get": {
        "operationId": "BillingController_export",
        "parameters": [
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "example": "2026-10-01",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "description": "Exclu",
            "schema": {
              "example": "2026-11-01",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Écritures comptables de la période (ventes, banque, caisse, OD)",
        "tags": [
          "Facturation"
        ]
      }
    },
    "/v1/settings/billing": {
      "get": {
        "operationId": "BillingController_getSettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Facturation"
        ]
      },
      "put": {
        "operationId": "BillingController_putSettings",
        "parameters": [
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BillingSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Facturation"
        ]
      }
    },
    "/v1/dashboard": {
      "get": {
        "operationId": "DashboardController_get",
        "parameters": [
          {
            "name": "date",
            "required": false,
            "in": "query",
            "description": "Jour affiché (défaut : aujourd’hui, fuseau de l’établissement)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accept-language",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tableau de bord du jour, adapté aux droits du profil",
        "tags": [
          "Reporting"
        ]
      }
    },
    "/v1/reports/kpis": {
      "get": {
        "operationId": "DashboardController_kpi",
        "parameters": [
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "example": "2026-10-01",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "description": "Exclu",
            "schema": {
              "example": "2026-11-01",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Occupation, ADR, RevPAR et durée moyenne de séjour sur une période",
        "tags": [
          "Reporting"
        ]
      }
    }
  },
  "info": {
    "title": "API – Système de gestion hôtelière",
    "description": "Socle V1 : session, menu dynamique, chambres, clients, disponibilités, réservations, réception.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ChargeDto": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "enum": [
              "restaurant",
              "bar",
              "minibar",
              "spa",
              "laundry",
              "phone",
              "other"
            ]
          },
          "description": {
            "type": "string",
            "example": "Dîner – menu du jour"
          },
          "quantity": {
            "type": "number",
            "default": 1
          },
          "unitAmount": {
            "type": "number",
            "example": 2800,
            "description": "Prix unitaire TTC en unités mineures"
          },
          "vatRate": {
            "type": "number",
            "example": 1000,
            "description": "Taux de TVA en points de base ; défaut selon la catégorie"
          }
        },
        "required": [
          "category",
          "description",
          "unitAmount"
        ]
      },
      "PaymentDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "deposit",
              "payment",
              "refund"
            ]
          },
          "method": {
            "type": "string",
            "enum": [
              "cash",
              "card",
              "transfer",
              "purchase_order"
            ]
          },
          "amount": {
            "type": "number",
            "example": 29000,
            "description": "Montant positif en unités mineures (un remboursement est enregistré en négatif)"
          },
          "pspReference": {
            "type": "string",
            "description": "Référence de transaction du prestataire de paiement (obligatoire pour la carte). Jamais le numéro de carte."
          },
          "reference": {
            "type": "string",
            "description": "Référence du virement ou numéro de bon de commande"
          }
        },
        "required": [
          "kind",
          "method",
          "amount"
        ]
      },
      "IssueInvoiceDto": {
        "type": "object",
        "properties": {
          "lang": {
            "type": "string",
            "enum": [
              "fr",
              "en",
              "ar"
            ],
            "description": "Langue de la facture (défaut : langue du profil)"
          }
        }
      },
      "CreditNoteDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "Erreur sur le nombre de nuits"
          }
        },
        "required": [
          "reason"
        ]
      },
      "VatDto": {
        "type": "object",
        "properties": {}
      },
      "TouristTaxDto": {
        "type": "object",
        "properties": {}
      },
      "SellerDto": {
        "type": "object",
        "properties": {}
      },
      "BillingSettingsDto": {
        "type": "object",
        "properties": {
          "country": {
            "type": "string",
            "example": "FR",
            "description": "Pays d’exploitation (ISO 3166-1)"
          },
          "vat": {
            "$ref": "#/components/schemas/VatDto"
          },
          "touristTax": {
            "$ref": "#/components/schemas/TouristTaxDto"
          },
          "seller": {
            "$ref": "#/components/schemas/SellerDto"
          },
          "accounts": {
            "type": "object",
            "description": "Plan de comptes de l’export comptable"
          }
        }
      }
    }
  }
}